Ditches one-shot-modifiers
This required finding a home for all the modifiers that didn't conflict
with the symbol layer, which I managed to do.

My experience with the one shot modifiers was
  a) The key detection was not quite reliable enough
  b) I was not accurate enough hitting Ctrl vs Cmd
  c) My brain couldn't handle "lock-mode"

I may experiment with it again at some point, possibly with
modifications to lock mode and key placement. For now I'm back to normal
modifiers.