Tweaks symbols: removes *some* tapdance, adds more symbols
After some months of use, some keys remained unused on the
symbol layer. This allowed me to break up the tap dances that
have been annoying (I still like the bracket tap dances) and
add a few more symbols to the layer that had never made it
there in the first place.

* Broke up DanceSlash and DanceQuotes into individual keys
* Added `, ?, ~ keys to the symbol layer.