Fixes OSM broken stickability control with newest OSM
It looks like OneShotModifiers added new functions for controlling
stickability, but left the old booleans intact. The old booleans
appear to not be functioning correctly, however. This updates this
firmware to use the new interface, which fixes the issue where
OneShotModifiers were stickable even when they should not have been.

I noticed the new interface to stickabality in this Kaleidoscope commit:
https://github.com/keyboardio/Kaleidoscope/commit/28a589be03db6827e89e24fe82141dae475baa30