Removes tapdance, sacrificing function keys to brackets
I have never gotten as proficient with tapdance as I hoped I would. In
addition, there is a bug with repeating keys that has been causing me
frequent issues when I accidentally hold down a tapdance key while
trying to mousekeys.

This is an experiment to see whether I prefer a world without tapdance
now that I have used it for quite some time. Originally I was focused on
keeping all the symbols on the 3 central rows, but now some of the
brackets are on the top row, replacing function key mappings that I
never use anyway.